Customer user unfolders (#14)
The `gotype.NewUnfolder` constructor now accepts functional options. Currently only the `Unfolders` option is supported. `Unfolders` accepts a number of primitive or processing unfolders. These are plain functions with different signature, depending on target type and required processing. See `TestUserUnfold` for sample use cases.

For rich structures the unfolder pre-computes and stores unfolders for compound structures. This cache used to be global/shared, and protected via a mutex. With the introduction of user unfolders the cache is not shared anymore, due to potential conflicts between type representations. Each instance of `gotype.Unfolder` no has a local cache only. The unfolder was not thread safe before. There is no change in semantics for library users.

**Primitive user unfolders**: Primitive unfolders are used for converting a primitive value into the target type. The type signature for a primitive unfolder must be:
`func(to *Target, from P) error`, where `to` is an arbitrary go type that the result should be written to and `P` must be one of: `bool, string, uint(8|16|32|64), int(8|16|32|64), float(32|64)`. The unfolder will be used for any value of type `Target` or `*Target`.
Unfolding fails if the value presented to the unfolder is not compatible with `P`.

**Processing user unfolder**: Processing unfolders provide an init function, that returns a memory cell for temporary parsing and a post processing function. `gotype` first parses the input according to the type of the memory cell. This way one can parse an array or object into a temporary value, and then compute the actual value matching the target type.
The signature of the `init` function is: `func(to *Target) (cell interface{}, after processFn)`. The init receives a pointer to the target value of type `Target`. The init function can pre-initialize `to` if required. It returns a temporary store for parsing purposes via `cell`. If the type/byte layout doesn't change, then one can return `*to` for cell. The `processFn` signature is: `func(to *Target, cell interface{}) error`. Both, the target pointer and cell are given to the processing function again (no closure to store intermediate state is required). If the `Target` type is not reachable from the `Target` type itself, then it is safe to allocate and reuse the `cell` between multiple calls. But each instance of `gotype.Unfolder` must have it's own preallocated `cell`.